shloka part done

+++ bins/models.py
@@ -9,14 +9,14 @@ WASTE_TYPES = (('cardboard','Cardboard'),
 	('plastic','Plastic'),
 	('trash','Trash'))
 class SmartBins(models.Model):
-	latitude = models.IntegerField()
-	longitude = models.IntegerField()
+	latitude = models.FloatField()
+	longitude = models.FloatField()
 	location_name = models.CharField(max_length=50)
 	garbage_value = models.FloatField()
 	waste_type = models.CharField(max_length=10,choices=WASTE_TYPES,default="cardboard")
 	needs_to_be_collected = models.BooleanField(default=False)
 	bin_full = models.BooleanField(default=False)
 	waste_collected = models.BooleanField(default=False)
-	
+
 	def __str__(self):
 		return self.location_name
